requirements.txt
python-telegram-bot==21.6
APScheduler==3.6.3
pytz
tzlocal
cachetools
tornado
setuptools

